Removing tex symbol to utf8 conversion from abstract
In arxiv-browse, the abs template was using tex_2_utf on the
abstract. This was converting TeX like \psi to the unicode φ. This is
bad because φ doesn't look like a TeX \psi.

The abstract on /abs now gets tex_to_utf_no_symb which doesn't
do symbols.

Also \psi is now mapped to ϕ instead of φ. \varphi is mapped to φ.

The classic system does not do tex_2_utf on the abstract. It only
does normal HTML escaping.

ARXIVNG-1612