Fixes abs parse error with backslashes
Changes how the .abs file is split between the top metadata and the
abstract. It was using a regex and now it does a loop over the lines.

I moved away from the regex to avoid bad interactions with the section
delimiters and LaTeX line breaks. Both of these are \\ (two
backslashes).

ARXIVNG-3976