Uses CloudPathLib in more places, gettz to ZoneInfo
Changes more places in the code to use CloudpPathLib. Including in the
prev/next code.

Move to pass timezones down into more functions, and to get the timezone from
the `app.config`. The calls to app.config no longer have a default since
if the default in app.config is not set, this things are in a messed up state.

Changed how imports are in `browse.services.document` to avoid circular imports.