Changes to parse_abs to support listings
These are changes to support the goal of reusing the abs parsing to
parse the new type listings files.

Splits the parse_abs into functions that handle different parts.

Adds RE_REP_COMPONENTS to handle date in version history that is found in
replacements of the new type listing files.