WIP Adds quick parameter to get_dissemination_formats()
Tests fail on older listing files.

The quick parameter will get the dissemination formats based only
on the info in the source_type. When not set to quick it will
also check the source file names.

The reason for this is to avoid looking up the src file names while
building long listing pages.

db_abs sets source_type to a SourceType instead of str.