Sets the cloudpathlib caching mode to close_file
Rational:

GCP CloudRun has as its file system as RAM FS so everything written to disk
actually ends up using memory.

cloudpathlib does not attempt to be thread safe if a single `Client` is used
across multiple threads. As of 2023-06 when it writes and deletes from its cache
there is no code to check for race conditions. Ex `Client.clear_cache()`.

Flask apps run multi threaded.

So we have two problems:

1. cloudpathlib is not thread safe but we must use threads
2. we'll fill up memory with cloudpathlib's default caching

The work already exists before this commit to have per thread `Client` instances
and this commit adds the cache file mode of "close_file". Which removes the file
from the cache when it is closed.