Fixes case a of 404 for an ancillary file
The ancillary files are served from .tar but for efficiency the
tar was not read until it is passed back to werkzeug as a generator.
With this design it was not until werkzeug was processing the stream that
the absence of the file from the tar was detected.

This changes it so the check for the absence happens before the response is
created for werkzeug.

Since the `tarinfo` object has the offsets into the tar, an exists
check can happen with out having to read the whole tar twice.