Changes the Cache-Control on /pdf no_source responses
Sets it to 3000 seconds and status 500 so it shows up better in the logs.

The cache control was set to the max value. This makes it so that the CDN caches
a PDF that is missing and the the CDN won't check again in months and months to
see if the PDF has been synced.

This was sending the no_source page. Because both the source and PDf are
missing.  Ensure that this sets a cache-control of just a couple minutes.

ARXIVCE-776