Changes behavior of /src
This changes /src/{paperid} to download the source in the format that it exists
in the storage. This is a behavior change that affects users.

The reasons for this change:

1. fastly and other CDNs need to have Range Requests and partial content enabled
to handle large files (> 20MB) and Content-Length

2. We want to use streaming to avoid copying large files to the limited storage
on containers.

3. The final Content-Length will not be known when streaming a response out via a pipe of tar -> gzip

Old behavior described below this line:
The path /src use to always download a .tar.gz file. It would always set the
encoding to gzipped.

In the common case of .tar.gz it just sent the file.

In the case of a single file like .pdf it tarred and gziped that.

In the case of a .gz single file like a .ps.gz it would unzip, tar and regzip.