Fixes pdf redirect to http bug ARXIVCE-2718
The bug was that a URL like https://arxiv.org/pdf/2501.12345.pdf would redirect
to http://arxiv.org/pdf/2501.12345. Note the https is chagned to http and the
trailing .pdf is removed. The chagne from https to http is the bug. This commit
fixes that bug.

Changes the abs HTML metadata pdf URL from a http to https.

Changes the author list page pdf URL from a http to https.